+++ .idea/encodings.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/sortAlgorithm/src/daeHwan/NaturalMergeSort.java" charset="US-ASCII" />
     <file url="file://$PROJECT_DIR$/sortAlgorithm/src/daeHwan/Test.java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/sortAlgorithm/src/sort/NaturalMergeSort.java" charset="US-ASCII" />
+    <file url="file://$PROJECT_DIR$/sortAlgorithm/src/sort/NaturalMergeSort.java" charset="US-ASCII" />
     <file url="PROJECT" charset="x-windows-949" />
   </component>
 </project>
