--- sortAlgorithm/src/sort/README.txt
@@ -1,11 +1,0 @@
-각 개인 작업공간에서 작업 후 sort 클래스들은 이 디랙토리로 옮겨서 완성하겠습니다
-리퀘스트 올리고 카톡주면 반영할텐데, 개인별 할당된 sort algorithm 만 구현하면 되어서 딱히 신경 쓸건 없을겁니다
-
-메인에서 여기있는 클래스들 끌어다가 사용해서 시간을 측정할 예정입니다
-
-
-sort 클래스가 구현해야 할 조건
-    - ClassName.sort(array) 로 array 정렬
-    - 각 SortClass 는 sort 메서드의 가인수를 (Comparable[] arr)를 받아오도록
-    - 입력배열은 빈 공간이 없음
-    - 가급적 간결하고 효율좋은 코드 사용 -> 인터넷발 코드나 강의자료 코드로, 자신이 구현한 것과 속도 비교하기
