+++ sortAlgorithm/src/sort/RecursiveQuickSort.java
@@ -1,6 +1,7 @@
-package jongHyeon;
+package sort;
 
-public class MedianOfThreeQuickSort {
+public class RecursiveQuickSort {
+
     public static void sort(Comparable[] a) {
         int low = 0;
         int high = a.length - 1;
@@ -14,10 +15,8 @@ public class MedianOfThreeQuickSort {
         sort(a, j + 1, high);
     }
 
-    public static int partition(Comparable[] a, int low, int high) {
-        int mid = (low + high) / 2;
-        setPivot(a, low, mid, high);
 
+    public static int partition(Comparable[] a, int low, int high) {
         int pivot = low;
         Comparable standard = a[pivot];
         int i = low + 1;
@@ -31,33 +30,9 @@ public class MedianOfThreeQuickSort {
         swap(a, pivot, j);
         return j;
     }
-
     public static void swap(Comparable[] a, int i, int j) {
         Comparable temp = a[i];
         a[i] = a[j];
         a[j] = temp;
-    }
-
-    public static void setPivot(Comparable[] a, int low, int mid, int high) {
-        if(a[low].compareTo(a[mid]) >= 0) {
-            if(a[mid].compareTo(a[high]) >= 0) {
-                swap(a, low, mid);
-            }
-            else if(a[low].compareTo(a[high]) <= 0) {
-
-            }
-            else {
-                swap(a, low, mid);
-            }
-        }
-        else if(a[high].compareTo(a[low]) < 0) {
-
-        }
-        else if(a[mid].compareTo(a[high]) > 0) {
-            swap(a, low, high);
-        }
-        else {
-            swap(a, low, mid);
-        }
     }
 }
