+++ sortAlgorithm/src/sort/ShellSort.java
@@ -1,6 +1,4 @@
-package minJun;
-
-import java.lang.Comparable;
+package sort;
 
 public class ShellSort {
     public static void sort(Comparable[] array) {
